rename: quantumion.interfaces to quantumion.interface
