+++ quantumion/backend/digital/python/tc.py
@@ -5,7 +5,7 @@ import tensorcircuit as tc
 from quantumion.backend.base import BackendBase
 from quantumion.backend.task import Task, TaskResultDigital
 
-from quantumion.interfaces.digital.circuit import DigitalCircuit
+from quantumion.interface.digital.circuit import DigitalCircuit
 
 ########################################################################################
