+++ quantumion/backend/task.py
@@ -6,10 +6,10 @@ from pydantic import BaseModel
 
 ########################################################################################
 
-from quantumion.interfaces.math import ComplexFloat
-from quantumion.interfaces.analog.circuit import AnalogCircuit
-from quantumion.interfaces.digital.circuit import DigitalCircuit
-from quantumion.interfaces.atomic.program import AtomicProgram
+from quantumion.interface.math import ComplexFloat
+from quantumion.interface.analog.circuit import AnalogCircuit
+from quantumion.interface.digital.circuit import DigitalCircuit
+from quantumion.interface.atomic.program import AtomicProgram
 
 from quantumion.backend.metric import Metric
