+++ quantumion/interface/math.py
@@ -2,7 +2,7 @@ import numpy as np
 
 ########################################################################################
 
-from quantumion.interfaces.base import VisitableBaseModel
+from quantumion.interface.base import VisitableBaseModel
 
 ########################################################################################
