+++ tests/test_analog/test_visitors.py
@@ -2,7 +2,7 @@ import numpy as np
 
 ########################################################################################
 
-from quantumion.interfaces.analog import *
+from quantumion.interface.analog import *
 from quantumion.compiler.analog import AnalogCircuitIonsAnalysis
 
 ########################################################################################
