+++ tests/test_atomic/test_visitors.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from quantumion.interfaces.atomic import (
+from quantumion.interface.atomic import (
     Level,
     Transition,
     Multipole,
