+++ tests/test_backends/test_qutip.py
@@ -4,7 +4,7 @@ from rich import print as pprint
 
 ########################################################################################
 
-from quantumion.interfaces.analog import (
+from quantumion.interface.analog import (
     PauliX,
     PauliY,
     PauliZ,
