+++ tests/test_serialization.py
@@ -6,8 +6,8 @@ import argparse
 
 ########################################################################################
 
-from quantumion.interfaces.analog import *
-from quantumion.interfaces.digital import *
+from quantumion.interface.analog import *
+from quantumion.interface.digital import *
 
 from quantumion.backend.task import Task, TaskArgsAnalog, TaskArgsDigital
