+++ tests/test_server.py
@@ -7,8 +7,8 @@ import argparse
 from quantumion.backend.task import Task, TaskArgsAnalog, TaskArgsDigital
 from quantumion.backend.provider import Provider
 
-from quantumion.interfaces.analog import *
-from quantumion.interfaces.digital import *
+from quantumion.interface.analog import *
+from quantumion.interface.digital import *
 
 ########################################################################################
